Replace `size_t` with `int` throughout, for speed.
2 billion rows and columns should be enough for anyone.

21775 grids per second.

Investigating all 2659328000 grids would still take 33.9 hours, though.

+++ dance.h
@@ -17,17 +17,17 @@ struct data_object {
 
 struct column_object {
     struct data_object data; /* must be the first field */
-    size_t size;
+    int size;
     int name;
 };
 
 struct dance_matrix {
-    size_t nrows, ncolumns;
+    int nrows, ncolumns;
     struct column_object *columns;
     struct column_object head;
 };
 
-void dance_init(struct dance_matrix *m, size_t rows, size_t cols, const int *data);
-void dance_addrow(struct dance_matrix *m, size_t nentries, size_t *entries);
+void dance_init(struct dance_matrix *m, int rows, int cols, const int *data);
+void dance_addrow(struct dance_matrix *m, int nentries, int *entries);
 void dance_free(struct dance_matrix *m);
-int dance_solve(struct dance_matrix *m, dance_result (*f)(size_t, struct data_object **));
+int dance_solve(struct dance_matrix *m, dance_result (*f)(int, struct data_object **));
