+++ sudoku.cc
@@ -1,13 +1,12 @@
 
 #include "sudoku.h"
 
-#include <stdbool.h>
 #include <stdio.h>
 #include "dance.h"
 
 static size_t g_sudoku_counter = 0;
 
-static dance_result count_sudoku_result(size_t n, struct data_object **cols)
+static dance_result count_sudoku_result(int n, struct data_object **cols)
 {
     (void)n;
     (void)cols;
@@ -19,13 +18,13 @@ static dance_result count_sudoku_result(size_t n, struct data_object **cols)
 
 static int solve_sudoku_with_callback(
     const int grid[9][9],
-    dance_result (*f)(size_t, struct data_object **))
+    dance_result (*f)(int, struct data_object **))
 {
     g_sudoku_counter = 0;
 
     struct dance_matrix mat;
     int ns;
-    size_t constraint[4];
+    int constraint[4];
     int rows = 0;
     int cols = 9*(9+9+9)+81;
     /*
@@ -89,11 +88,11 @@ void print_sudoku_grid(const int grid[9][9])
     }
 }
 
-static dance_result print_unique_sudoku_result(size_t n, struct data_object **sol)
+static dance_result print_unique_sudoku_result(int n, struct data_object **sol)
 {
     int grid[9][9];
 
-    for (size_t i=0; i < n; ++i) {
+    for (int i=0; i < n; ++i) {
         int constraint[4];
         int row, col, val;
         row = col = val = 0;  /* shut up "unused" warning from compiler */
@@ -101,7 +100,7 @@ static dance_result print_unique_sudoku_result(size_t n, struct data_object **so
         constraint[1] = sol[i]->column->name;
         constraint[2] = sol[i]->right->column->name;
         constraint[3] = sol[i]->right->right->column->name;
-        for (size_t j=0; j < 4; ++j) {
+        for (int j=0; j < 4; ++j) {
             if (constraint[j] < 81) {
                 row = constraint[j] / 9;
                 val = constraint[j] % 9 + 1;
